organizer apis update for verification status from sa

+++ controllers/organizer/offers/online/remove.js
@@ -8,7 +8,7 @@ exports.remove = async (req, res) => {
     if (req.token.organizerid && mongoose.Types.ObjectId.isValid(req.token.organizerid)) {
         let primary = mongoConnection.useDb(constants.DEFAULT_DB);
         let organizerData = await primary.model(constants.MODELS.organizers, organizerModel).findById(req.token.organizerid).select('-password').lean();
-        if (organizerData && organizerData.status == true && organizerData.mobileverified == true) {
+        if (organizerData && organizerData.status == true && organizerData.mobileverified == true && organizerData.is_approved == true) {
             const { onlineofferid } = req.body;
             if (onlineofferid && onlineofferid != '' && mongoose.Types.ObjectId.isValid(onlineofferid)) {
                 await primary.model(constants.MODELS.onlineoffers, onlineofferModel).findByIdAndRemove(onlineofferid);
